split up tests to functions as there are too many cases to test and it adds up so quickly
